% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadthat}
\alias{downloadthat}
\title{Download file/webpage}
\usage{
downloadthat(.x, .o)
}
\arguments{
\item{.x}{Path to file or URL}

\item{.o}{File name to save the text/source as.}
}
\value{
A string of read-in text
}
\description{
Downloads text/source from file or URL
}
