% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-those.R
\name{downloadthose}
\alias{downloadthose}
\title{Download file(s)/webpage(s)}
\usage{
downloadthose(.x, .o)
}
\arguments{
\item{.x}{Paths to files or URLs}

\item{.o}{File names to save the texts/sources as.}
}
\value{
A character vector where each element is a string of read-in text
  associated with each input.
}
\description{
Downloads text/source from multiple files or URLs
}
