% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readit-package.R
\docType{package}
\name{readit-package}
\alias{readit}
\alias{readit-package}
\title{readit: Read Text Data}
\description{
Quickly read in text from files or URLs.
}
\author{
\strong{Maintainer}: Michael W. Kearney \email{kearneymw@missouri.edu} (0000-0002-0730-4694)

}
\keyword{internal}
