% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/that-read.R
\name{readthat}
\alias{readthat}
\alias{readthat2}
\title{Read file/webpage}
\usage{
readthat(.x)

readthat2(.x)
}
\arguments{
\item{.x}{Path to file or URL}
}
\value{
A string of read-in text
}
\description{
Returns text/source from file or URL
}
