% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readthose}
\alias{readthose}
\title{Read file(s)/webpage(s)}
\usage{
readthose(.x)
}
\arguments{
\item{.x}{Paths to files or URLs}
}
\value{
A character vector where each element is a string of read-in text
  associated with each input.
}
\description{
Returns text/source from multiple files or URLs
}
