% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readthat-package.R
\docType{package}
\name{readthat-package}
\alias{readthat}
\alias{readthat-package}
\title{readthat: Read Text Data}
\description{
Quickly read in text from files or URLs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mkearney/readthat}
  \item Report bugs at \url{https://github.com/mkearney/readthat/issues}
}

}
\author{
\strong{Maintainer}: Michael W. Kearney \email{kearneymw@missouri.edu} (0000-0002-0730-4694)

}
\keyword{internal}
