% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readthat-package.R
\name{readthat}
\alias{readthat}
\title{Read that file/webpage}
\usage{
readthat(.x)
}
\arguments{
\item{.x}{Path to file or URL}
}
\value{
A string of read-in text
}
\description{
Returns text/source from file or URL
}
